connect forecast api, create and append to html

+++ script.js
@@ -10,6 +10,7 @@ var APIKKey = "4e82548e9a516fc5ec14f1ab6d4a2c48"
 
 
 // function to get weather data from openweathermap api when user searches for a city
+
 var formSubmitHandler = function(event) {
     event.preventDefault();
 
@@ -18,18 +19,21 @@ var formSubmitHandler = function(event) {
 
     // if cityname exists, get weather data for that city
     if (cityname) {
+                // check if cityname already exists in recent search buttons
+                var existingButton = document.querySelector(`button[data-cityname="${cityname}"]`);
+                if (!existingButton) {
+                    // append data attributes to recent search buttons
+                    // save cityname to localStorage
+                    localStorage.setItem('cityname', cityname);
+                    // create button for recent search and set data attribute to cityname and text to cityname with a class of btn
+                    var recentButton = document.createElement('button');
+                    recentButton.setAttribute('data-cityname', cityname);
+                    recentButton.textContent = cityname;
+                    recentButton.classList.add('btn');
+                    recentButtonsEl.appendChild(recentButton);
+                }
         getCityWeather(cityname);
-        //append data attributes to recent search buttons
-        // save cityname to localStorage
-        localStorage.setItem('cityname', cityname);
-        // create button for recent search and set data attribute to cityname and text to cityname with a class of btn
-        var recentButton = document.createElement('button');
-        recentButton.setAttribute('data-cityname', cityname);
-        recentButton.textContent = cityname;
-        recentButton.classList.add('btn');
-        recentButtonsEl.appendChild(recentButton);
-        
-        // clear old content
+
         currentweatherContainerEl.textContent = '';
         cityInputEl.value = '';
     } else {
@@ -68,6 +72,27 @@ var getCityWeather = function(cityname) {
             console.log(data.main.temp);
             // call on displayWeather function
             displayWeather(data, cityname);
+            // get latitude and longitude data from response
+            var lat = data.coord.lat;
+            var lon = data.coord.lon;
+
+            getFiveDayForecast(lat, lon, cityname);
+        });
+};
+
+// function to get 5 day forecast data from openweathermap api
+getFiveDayForecast = function(lat, lon, cityname) {
+    // format the openweathermap api url
+    var queryURL = "https://api.openweathermap.org/data/2.5/forecast?lat=" + lat + "&lon=" + lon + "&units=imperial&appid=" + APIKKey;
+
+    // make a request to the url
+    fetch(queryURL)
+        .then(response => response.json())
+        .then(data => {
+            console.log(data);
+            console.log(queryURL);
+            // call on displayWeather function
+            displayFiveDayForecastWeather(data, cityname);
         });
 };
 
@@ -97,7 +122,58 @@ var displayWeather = function(data, cityname) {
     weatherContainer.appendChild(weatherReportEl);
 };
 
+var displayFiveDayForecastWeather = function(data, cityname) {
+    // create a div element to hold weather data
+    var weatherForecastContainer = document.getElementById('forecast-container');
+    weatherForecastContainer.innerHTML = '';
 
+
+    var dailyForecast = [];
+    //loop through each day in the forecast
+    for (var i = 0; i < data.list.length; i++) {
+        //get the data for the current day
+        var dateString = data.list[i].dt_txt;
+        
+
+        //create a new date object
+        var forecastdate = new Date(dateString);
+
+        // check if current data point is at 12pm
+        if (forecastdate.getHours() === 12) {
+            //create a new card element for each day
+            var ForecastCard = document.createElement('div');
+            ForecastCard.classList = 'forecast-card', 'card', 'col-12 col-md-7';
+
+            // add data to daily forecast array
+            dailyForecast.push(data.list[i]);
+            console.log(dailyForecast);
+
+
+
+
+    // create a span element to hold date
+    var forecastdateEl = document.createElement('span');
+    var dateObj = new Date(data.list[i].dt_txt); // convert to date object
+    var dateString = dateObj.toLocaleDateString(); // convert to localized date string
+    forecastdateEl.textContent = dateString;
+    forecastdateEl.classList = 'card-title';
+    // create an image element to hold weather icon
+    var weatherIconEl = document.createElement('img');
+    weatherIconEl.setAttribute('src', 'http://openweathermap.org/img/w/' + data.list[0].weather[0].icon + '.png');
+    // create a span element to hold temperature data
+    var temperatureEl = document.createElement('span');
+    temperatureEl.textContent = 'Temperature: ' + data.list[i].main.temp + ' °F';
+    temperatureEl.classList = 'card-text';
+
+    // append the cityname, weather icon, and temperature to the weather report div
+    var weatherForecastContainer = document.getElementById('forecast-container');
+    ForecastCard.appendChild(forecastdateEl);
+    ForecastCard.appendChild(weatherIconEl);
+    ForecastCard.appendChild(temperatureEl);
+    weatherForecastContainer.appendChild(ForecastCard);
+};
+};
+};
